test: fix test cases
